# contracts/openapi/knowledge-v1.yaml
openapi: 3.1.0
info: {title: AliAgent Knowledge API, version: 1.0.0}
paths:
  /api/v1/knowledge/documents:
    post:
      summary: 上传文件并创建异步摄入任务
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Subject'}, {$ref: '#/components/parameters/Trace'}, {$ref: '#/components/parameters/SubjectType'}, {$ref: '#/components/parameters/Roles'}, {$ref: '#/components/parameters/Permissions'}, {$ref: '#/components/parameters/AuthorizationSnapshot'}]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary}}}}}}
      responses: {'202': {description: 已接受}}
  /api/v1/knowledge/ingestion-tasks/{taskId}:
    get:
      summary: 查询摄入状态
      parameters: [{$ref: '#/components/parameters/Tenant'}, {name: taskId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: 任务状态}}
  /api/v1/knowledge/versions/{versionId}/publish:
    post:
      summary: 发布审核就绪版本
      parameters: [{$ref: '#/components/parameters/Tenant'}, {name: versionId, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: 已发布}, '409': {description: 非审核就绪状态}}
  /api/v1/knowledge/retrieval:query:
    post:
      summary: P3-B 检索契约，仅定义不在 P3-A 实现
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Subject'}, {$ref: '#/components/parameters/SubjectType'}, {$ref: '#/components/parameters/Roles'}, {$ref: '#/components/parameters/Permissions'}, {$ref: '#/components/parameters/AuthorizationSnapshot'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RetrievalRequest'}}}}
      responses: {'200': {description: 已发布版本的检索结果, content: {application/json: {schema: {$ref: '#/components/schemas/RetrievalResponse'}}}}}
components:
  parameters:
    Tenant: {name: X-Tenant-Id, in: header, required: true, schema: {type: string}}
    Subject: {name: X-Subject-Id, in: header, required: true, schema: {type: string}}
    Trace: {name: X-Trace-Id, in: header, required: true, schema: {type: string}}
    SubjectType: {name: X-Subject-Type, in: header, required: true, schema: {type: string, enum: [MEMBER, STAFF]}}
    Roles: {name: X-User-Roles, in: header, required: true, schema: {type: string}}
    Permissions: {name: X-User-Permissions, in: header, required: true, schema: {type: string}}
    AuthorizationSnapshot: {name: X-Authorization-Snapshot-Id, in: header, required: true, schema: {type: string, format: uuid}}
  schemas:
    RetrievalRequest:
      type: object
      required: [query]
      properties: {query: {type: string, minLength: 1}, topK: {type: integer, minimum: 1, maximum: 100, default: 10}}
    RetrievalResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {const: 200}
        message: {type: string}
        data: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/RetrievalItem'}}}}
    RetrievalItem:
      type: object
      required: [documentId, versionId, chunkId, content, score]
      properties:
        documentId: {type: string, format: uuid}
        versionId: {type: string, format: uuid}
        chunkId: {type: string, format: uuid}
        content: {type: string}
        score: {type: number}